+++ src/Analyzers/CSharp/Analyzers/UseCollectionExpression/UseCollectionExpressionHelpers.cs
@@ -96,7 +96,11 @@ internal static class UseCollectionExpressionHelpers
 
         var parent = topMostExpression.GetRequiredParent();
 
-        if (!IsInTargetTypedLocation(semanticModel, topMostExpression, cancellationToken))
+        if (!topMostExpression.IsInTargetTypedLocation(semanticModel, cancellationToken))
+            return false;
+
+        // (X[])[1, 2, 3] is target typed.  `(X)[1, 2, 3]` is currently not (because it looks like indexing into an expr).
+        if (topMostExpression.Parent is CastExpressionSyntax castExpression && castExpression.Type is IdentifierNameSyntax)
             return false;
 
         // X[] = new Y[] { 1, 2, 3 }
@@ -613,118 +617,6 @@ internal static class UseCollectionExpressionHelpers
                semanticModel.GetTypeInfo(expression, cancellationToken).Type?.IsValueType == true;
     }
 
-    private static bool IsInTargetTypedLocation(SemanticModel semanticModel, ExpressionSyntax expression, CancellationToken cancellationToken)
-    {
-        var topExpression = expression.WalkUpParentheses();
-        var parent = topExpression.Parent;
-        return parent switch
-        {
-            EqualsValueClauseSyntax equalsValue => IsInTargetTypedEqualsValueClause(equalsValue),
-            CastExpressionSyntax castExpression => IsInTargetTypedCastExpression(castExpression),
-            // a ? [1, 2, 3] : ...  is target typed if either the other side is *not* a collection,
-            // or the entire ternary is target typed itself.
-            ConditionalExpressionSyntax conditionalExpression => IsInTargetTypedConditionalExpression(conditionalExpression, topExpression),
-            // Similar rules for switches.
-            SwitchExpressionArmSyntax switchExpressionArm => IsInTargetTypedSwitchExpressionArm(switchExpressionArm),
-            InitializerExpressionSyntax initializerExpression => IsInTargetTypedInitializerExpression(initializerExpression, topExpression),
-            CollectionElementSyntax collectionElement => IsInTargetTypedCollectionElement(collectionElement),
-            AssignmentExpressionSyntax assignmentExpression => IsInTargetTypedAssignmentExpression(assignmentExpression, topExpression),
-            BinaryExpressionSyntax binaryExpression => IsInTargetTypedBinaryExpression(binaryExpression, topExpression),
-            LambdaExpressionSyntax lambda => IsInTargetTypedLambdaExpression(lambda, topExpression),
-            ArgumentSyntax or AttributeArgumentSyntax => true,
-            ReturnStatementSyntax => true,
-            ArrowExpressionClauseSyntax => true,
-            _ => false,
-        };
-
-        bool HasType(ExpressionSyntax expression)
-            => semanticModel.GetTypeInfo(expression, cancellationToken).Type is not null and not IErrorTypeSymbol;
-
-        static bool IsInTargetTypedEqualsValueClause(EqualsValueClauseSyntax equalsValue)
-            // If we're after an `x = ...` and it's not `var x`, this is target typed.
-            => equalsValue.Parent is not VariableDeclaratorSyntax { Parent: VariableDeclarationSyntax { Type.IsVar: true } };
-
-        static bool IsInTargetTypedCastExpression(CastExpressionSyntax castExpression)
-            // (X[])[1, 2, 3] is target typed.  `(X)[1, 2, 3]` is currently not (because it looks like indexing into an expr).
-            => castExpression.Type is not IdentifierNameSyntax;
-
-        bool IsInTargetTypedConditionalExpression(ConditionalExpressionSyntax conditionalExpression, ExpressionSyntax expression)
-        {
-            if (conditionalExpression.WhenTrue == expression)
-                return HasType(conditionalExpression.WhenFalse) || IsInTargetTypedLocation(semanticModel, conditionalExpression, cancellationToken);
-            else if (conditionalExpression.WhenFalse == expression)
-                return HasType(conditionalExpression.WhenTrue) || IsInTargetTypedLocation(semanticModel, conditionalExpression, cancellationToken);
-            else
-                return false;
-        }
-
-        bool IsInTargetTypedLambdaExpression(LambdaExpressionSyntax lambda, ExpressionSyntax expression)
-            => lambda.ExpressionBody == expression && IsInTargetTypedLocation(semanticModel, lambda, cancellationToken);
-
-        bool IsInTargetTypedSwitchExpressionArm(SwitchExpressionArmSyntax switchExpressionArm)
-        {
-            var switchExpression = (SwitchExpressionSyntax)switchExpressionArm.GetRequiredParent();
-
-            // check if any other arm has a type that this would be target typed against.
-            foreach (var arm in switchExpression.Arms)
-            {
-                if (arm != switchExpressionArm && HasType(arm.Expression))
-                    return true;
-            }
-
-            // All arms do not have a type, this is target typed if the switch itself is target typed.
-            return IsInTargetTypedLocation(semanticModel, switchExpression, cancellationToken);
-        }
-
-        bool IsInTargetTypedCollectionElement(CollectionElementSyntax collectionElement)
-        {
-            // We do not currently target type spread expressions in a collection expression.
-            if (collectionElement is not ExpressionElementSyntax)
-                return false;
-
-            // The element it target typed if the parent collection is itself target typed.
-            var collectionExpression = (CollectionExpressionSyntax)collectionElement.GetRequiredParent();
-            return IsInTargetTypedLocation(semanticModel, collectionExpression, cancellationToken);
-        }
-
-        bool IsInTargetTypedInitializerExpression(InitializerExpressionSyntax initializerExpression, ExpressionSyntax expression)
-        {
-            // new X[] { [1, 2, 3] }.  Elements are target typed by array type.
-            if (initializerExpression.Parent is ArrayCreationExpressionSyntax)
-                return true;
-
-            // new [] { [1, 2, 3], ... }.  Elements are target typed if there's another element with real type.
-            if (initializerExpression.Parent is ImplicitArrayCreationExpressionSyntax)
-            {
-                foreach (var sibling in initializerExpression.Expressions)
-                {
-                    if (sibling != expression && HasType(sibling))
-                        return true;
-                }
-            }
-
-            // TODO: Handle these.
-            if (initializerExpression.Parent is StackAllocArrayCreationExpressionSyntax or ImplicitStackAllocArrayCreationExpressionSyntax)
-                return false;
-
-            // T[] x = [1, 2, 3];
-            if (initializerExpression.Parent is EqualsValueClauseSyntax)
-                return true;
-
-            return false;
-        }
-
-        bool IsInTargetTypedAssignmentExpression(AssignmentExpressionSyntax assignmentExpression, ExpressionSyntax expression)
-        {
-            return expression == assignmentExpression.Right && HasType(assignmentExpression.Left);
-        }
-
-        bool IsInTargetTypedBinaryExpression(BinaryExpressionSyntax binaryExpression, ExpressionSyntax expression)
-        {
-            return binaryExpression.Kind() == SyntaxKind.CoalesceExpression && binaryExpression.Right == expression && HasType(binaryExpression.Left);
-        }
-    }
-
     public static CollectionExpressionSyntax ConvertInitializerToCollectionExpression(
         InitializerExpressionSyntax initializer, bool wasOnSingleLine)
     {
@@ -1076,13 +968,13 @@ internal static class UseCollectionExpressionHelpers
                 if (arguments.Count == 1 &&
                     compilation.SupportsRuntimeCapability(RuntimeCapability.InlineArrayTypes) &&
                     originalCreateMethod.Parameters is [
+                    {
+                        Type: INamedTypeSymbol
                         {
-                            Type: INamedTypeSymbol
-                            {
-                                Name: nameof(Span<int>) or nameof(ReadOnlySpan<int>),
-                                TypeArguments: [ITypeParameterSymbol { TypeParameterKind: TypeParameterKind.Method }]
-                            } spanType
-                        }])
+                            Name: nameof(Span<int>) or nameof(ReadOnlySpan<int>),
+                            TypeArguments: [ITypeParameterSymbol { TypeParameterKind: TypeParameterKind.Method }]
+                        } spanType
+                    }])
                 {
                     if (spanType.OriginalDefinition.Equals(compilation.SpanOfTType()) ||
                         spanType.OriginalDefinition.Equals(compilation.ReadOnlySpanOfTType()))
